[apple/mac] Lots of tuning work around apple export

+++ ui/Templates/Xcode4/iOS Template/iOS Template/ViewController.m
@@ -170,8 +170,7 @@ NSMutableArray *tableData;
      return result;
      else
      return UIInterfaceOrientationMaskAll;*/
-    return gdr_supportedInterfaceOrientations();
-
+    return (NSUInteger)gdr_supportedInterfaceOrientations();
 }
 
 - (void)initTable{
